fixed bug change to support float number

+++ question3/src/main/java/app/practice/question3/utils/NumberFormat.kt
@@ -7,7 +7,7 @@ val decFormat = DecimalFormat("#.#")
 
 fun String?.toDecimalFormat(): String {
     try {
-        val number = this?.toInt()
+        val number = this?.toFloat()
         return decFormat.format(number)
 
     } catch (e: Exception) {
